release/1.0.0

+++ indexing-service/src/main/java/com/itmo/indexing_service/model/transcation/gas/GasData.java
@@ -17,8 +17,8 @@ public class GasData {
     private String owner;
 
     @JsonProperty("price")
-    private BigInteger price;
+    private Long price;
 
     @JsonProperty("budget")
-    private BigInteger budget;
+    private Long budget;
 }
